Guess the Number
Hey there, this simple program is made in Python language. This program involves computer choosing a random number between two limits and the user tries to guess it. In case of wrong guess, the computer will give you hints like the number is too low or too high. This really simple program involves taking input from the user and then matching it with the random number that computer chose.