Rock Paper Scissor
Hello, You must have played rock, paper and scissor game in your childhood. Well, Here is the code version of that game. This really simple program involves taking input from the user and the computer. Then it interprets both the values. The result will be shown on the screen. This game has the same rules which are rock beats scissors, scissors cut paper and paper covers rock.